Move storage errors into separate file
